// packages/core/src/hydration/interfaces.d.ts
export declare const NUM_ROOT_NODES = "r";
export declare const TEMPLATE_ID = "i";
export declare const NODES = "n";
export declare const DISCONNECTED_NODES = "d";
export declare const I18N_DATA = "l";
export declare const DEFER_BLOCK_ID = "di";
export declare const DEFER_BLOCK_STATE = "s";
export declare const DEFER_PARENT_BLOCK_ID = "p";
export declare const DEFER_HYDRATE_TRIGGERS = "t";
export declare const DEFER_PREFETCH_TRIGGERS = "pt";
// ...
export interface SerializedDeferBlock {
    /**
     * This contains the unique id of this defer block's parent, if it exists.
     */
    [DEFER_PARENT_BLOCK_ID]?: string;
    /**
     * This field represents a status, based on the `DeferBlockState` enum.
     */
    [DEFER_BLOCK_STATE]?: number;
    /**
     * Number of root nodes that belong to this defer block's template.
     * This information is needed to effectively traverse the DOM tree
     * and add jsaction attributes to root nodes appropriately for
     * incremental hydration.
     */
    [NUM_ROOT_NODES]: number;
    /**
     * The list of triggers that exist for incremental hydration, based on the
     * `Trigger` enum.
     */
    [DEFER_HYDRATE_TRIGGERS]?: (DeferBlockTrigger | SerializedTriggerDetails)[];
}
export interface SerializedTriggerDetails {
    trigger: DeferBlockTrigger;
    delay?: number;
}